requirements.txt
# SoRa Core Requirements
# Clean extraction focused on core video generation functionality

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP and Async
httpx>=0.25.0
aiofiles>=23.2.1

# Database and Storage
psycopg2-binary>=2.9.7  # PostgreSQL
redis>=5.0.0

# Vector Store (Optional - will gracefully fallback if not installed)
chromadb>=0.4.15
# supabase>=2.0.0

# AI and ML (Optional - will use mock responses if not installed)
# transformers>=4.35.0
# torch>=2.1.0
# torchvision>=0.16.0
# sentence-transformers>=2.2.2

# Google AI (for CINEGEN storytelling)
google-generativeai>=0.3.0

# Image Processing
Pillow>=10.1.0

# Cloud Storage (Optional)
# boto3>=1.34.0  # For S3 storage

# Background Tasks
# celery>=5.3.0  # Optional for advanced job processing

# Utilities
python-dateutil>=2.8.0
python-slugify>=8.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0